Renamed solved to resolved and updated some notebooks to use the new op
