+++ proto/core/node/common/service/resolve_pointer_type_service.proto
@@ -5,14 +5,14 @@ package syft.core.node.common.action;
 import "proto/core/common/common_object.proto";
 import "proto/core/io/address.proto";
 
-message SolvePointerTypeMessage {
+message ResolvePointerTypeMessage {
   syft.core.common.UID id_at_location = 1;
   syft.core.io.Address address = 2;
   syft.core.io.Address reply_to = 3;
   syft.core.common.UID msg_id = 4;
 }
 
-message SolvePointerTypeAnswerMessage {
+message ResolvePointerTypeAnswerMessage {
   string type_path = 1;
   syft.core.io.Address address = 2;
   syft.core.common.UID msg_id = 3;
