+++ src/syft/core/node/common/node.py
@@ -64,7 +64,7 @@ from .service.obj_search_permission_service import (
 )
 from .service.obj_search_service import ImmediateObjectSearchService
 from .service.repr_service import ReprService
-from .service.solve_pointer_type_service import SolvePointerTypeService
+from .service.resolve_pointer_type_service import ResolvePointerTypeService
 
 # this generic type for Client bound by Client
 ClientT = TypeVar("ClientT", bound=Client)
@@ -205,7 +205,7 @@ class Node(AbstractNode):
         self.immediate_services_with_reply: List[Any] = []
         self.immediate_services_with_reply.append(ImmediateObjectActionServiceWithReply)
         self.immediate_services_with_reply.append(ImmediateObjectSearchService)
-        self.immediate_services_with_reply.append(SolvePointerTypeService)
+        self.immediate_services_with_reply.append(ResolvePointerTypeService)
 
         # for services which can run at a later time and do not return a reply
         self.eventual_services_without_reply = list()
